% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account_status.R
\name{account_status}
\alias{account_status}
\title{Account Status}
\usage{
account_status()
}
\description{
This Function Returns Account Status.
}
\examples{
account_status()
}
\keyword{status}
