% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asks.R
\name{asks}
\alias{asks}
\title{Asks}
\usage{
asks(book, aggregated)
}
\arguments{
\item{book}{As String}
}
\description{
This Function Returns Active Ask Orders.
}
\examples{
asks("btc_mxn",TRUE)
}
\seealso{
\code{\link{get_books}}
}
\keyword{asks}
