% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_books.R
\name{available_books}
\alias{available_books}
\title{Available Books}
\usage{
available_books()
}
\description{
Currencies Available for Trade.
}
\examples{
available_books()

}
\keyword{books,currency}
