% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bids.R
\name{bids}
\alias{bids}
\title{Bids}
\usage{
bids(book, aggregated)
}
\arguments{
\item{book, }{Character}

\item{aggregated, }{Boolean}
}
\description{
This Function Returns Active Bids.
}
\examples{
bids()
}
\keyword{bids}
