% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cancel_order.R
\name{cancel_order}
\alias{cancel_order}
\title{Cancel Order}
\usage{
cancel_order(oid)
}
\arguments{
\item{oid, }{Character with order Id.}
}
\description{
This Function Deletes Specified Order.
}
\examples{
cancel_order("001Hg2risFrBul2J4G")
}
\keyword{cancel}
