% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funding_destination.R
\name{funding_destination}
\alias{funding_destination}
\title{Funding Destination}
\usage{
funding_destination(currency)
}
\arguments{
\item{currency, }{Character}
}
\description{
This Function Returns currency Funding Destination.
}
\examples{
funding_destination("mxn")
}
\keyword{funding,destination}
