% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fundings.R
\name{fundings}
\alias{fundings}
\title{Fundings}
\usage{
fundings(fid, limit)
}
\arguments{
\item{fid, }{Character, funding Id (optional).}

\item{limit, }{Numeric, history limit (optional).}
}
\description{
This Function Returns Funding History.
}
\examples{
fundings(limit=5)
}
\keyword{funding}
