% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_orders.R
\name{open_orders}
\alias{open_orders}
\title{Open Orders}
\usage{
open_orders(book)
}
\arguments{
\item{book, }{Character.}
}
\description{
Returns List of User's Open Orders for Specific Book.
}
\examples{
open_orders("btc_mxn")
}
\keyword{orders,open}
