% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_trades.R
\name{order_trades}
\alias{order_trades}
\title{Order Trades}
\usage{
order_trades(oid)
}
\arguments{
\item{oid, }{character: order identification number}
}
\description{
Returns Listed Trades Generated By Open Order
}
\examples{
order_trades("00K4kdad22Jaoij4Kahsfd")
}
\keyword{order,trades}
