% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orders.R
\name{orders}
\alias{orders}
\title{Orders}
\usage{
orders(oid)
}
\arguments{
\item{oid, }{character: order identification number (optional)}
}
\description{
Returns Open Order(s) Details.
}
\examples{
orders()
}
\keyword{orders,open}
