% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/place_order.R
\name{place_order}
\alias{place_order}
\title{Place Order}
\usage{
place_order(book, side, major, price, type)
}
\arguments{
\item{book, }{character: "mayor_minor"}

\item{side, }{character:  trading position, "buy" / "sell"}

\item{major, }{character: amount to trade on major units}

\item{price, }{character: proposed major price on minor units}

\item{type, }{character: type of order, "limit"}
}
\description{
Place Trading Order.
}
\examples{
order("btc_mxn","sell","0.001","75000","limit")
}
\keyword{order,sell,buy}
