% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ticker.R
\name{ticker}
\alias{ticker}
\title{Ticker}
\usage{
ticker(book)
}
\arguments{
\item{book, }{character}
}
\description{
Returns Detailed Book Ticker.
}
\examples{
ticker("btc_mxn")
}
\keyword{ticker,price,last,max,min,value,high,low}
