% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_parse.R
\name{time_parse}
\alias{time_parse}
\title{Time Parse}
\usage{
time_parse(tick_time)
}
\arguments{
\item{tick_time}{Character: Date-Time object}
}
\value{

}
\description{
Bitso API Time-Stamp Convertion.
}
\examples{
time_parse("2017-09-02T17:23:06+00:00")

}
