% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trades.R
\name{trades}
\alias{trades}
\title{Trades}
\usage{
trades(book)
}
\arguments{
\item{book, }{character}
}
\description{
Public Book Trading History.
}
\examples{
trades("btc_mxn")
}
\keyword{fees}
