% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withdrawals.R
\name{withdrawals}
\alias{withdrawals}
\title{Withdrawals}
\usage{
withdrawals(wid, limit)
}
\arguments{
\item{wid, }{character, withdrawal identification number}

\item{limit, }{numeric, number of entries to display}
}
\description{
User's Withdrawaling History.
}
\examples{
withdrawals()
}
\keyword{withdrawals}
